test(VerifierTest): Verifier 관련 테스트 패키지 이동
